# Updated sha libraries