--- a/src/uu/hashsum/Cargo.toml
+++ b/src/uu/hashsum/Cargo.toml
@@ -15,7 +15,7 @@ edition = "2018"
 path = "src/hashsum.rs"
 
 [dependencies]
-digest = "0.6.1"
+digest = "0.10.1"
 clap = { version = "3.0", features = ["wrap_help", "cargo"] }
 hex = "0.2.0"
 libc = "0.2.42"
@@ -24,8 +24,8 @@ md5 = "0.3.5"
 regex = "1.0.1"
 regex-syntax = "0.6.7"
 sha1 = "0.6.0"
-sha2 = "0.6.0"
-sha3 = "0.6.0"
+sha2 = "0.10.1"
+sha3 = "0.10.0"
 blake2b_simd = "0.5.11"
 uucore = { version=">=0.0.11", package="uucore", path="../../uucore" }