This gives a speedup of ~5x for some algorithms:

Old version:
```bash
$ time target/release/coreutils sha256sum ~/Downloads/ideaIC-2021.3.tar.gz 
de44097f00f4f9c48cf5abc2fab58c6436a4345f00f83ae7a75734af177e3077  /home/master/Downloads/ideaIC-2021.3.tar.gz

real	0m2,669s
user	0m2,576s
sys	0m0,092s
```

New version:
```bash
$ time target/release/coreutils sha256sum ~/Downloads/ideaIC-2021.3.tar.gz
de44097f00f4f9c48cf5abc2fab58c6436a4345f00f83ae7a75734af177e3077  /home/master/Downloads/ideaIC-2021.3.tar.gz

real	0m0,420s
user	0m0,348s
sys	0m0,072s
```