@sylvestre 2022-01-30
sweet
did you try with hyperfine (and compare with GNU's?)