@douglaz 2022-01-30
> hyperfine

```
Benchmark 1: target/release/coreutils sha256sum ~/Downloads/ideaIC-2021.3.tar.gz
  Time (mean ± σ):     430.3 ms ±   2.5 ms    [User: 353.6 ms, System: 76.6 ms]
  Range (min … max):   425.5 ms … 434.9 ms    10 runs
 
Benchmark 2: sha256sum ~/Downloads/ideaIC-2021.3.tar.gz
  Time (mean ± σ):      2.175 s ±  0.021 s    [User: 2.104 s, System: 0.069 s]
  Range (min … max):    2.132 s …  2.203 s    10 runs
 
Summary
  'target/release/coreutils sha256sum ~/Downloads/ideaIC-2021.3.tar.gz' ran
    5.05 ± 0.06 times faster than 'sha256sum ~/Downloads/ideaIC-2021.3.tar.gz'

```