@tertsdiepraam 2022-01-30
I can't replicate those improvements, sadly. Maybe they are platform-specific?