@douglaz 2022-01-30
> I can't replicate those improvements, sadly. Maybe they are platform-specific?

Tested on Ubuntu 20.04, AMD Ryzen 9 5900X cpu