@tertsdiepraam 2022-01-31
Hmm I guess the difference could be that I'm on intel? In any case, thanks for the PR!