@douglaz 2022-01-31
> Hmm I guess the difference could be that I'm on intel? In any case, thanks for the PR!

I tested it on a i3-10100 and got no improvement, so it seems to be a AMD-only optimization