@sylvestre 2022-01-31
Yeah, I can't replicate either on a Intel(R) Core(TM) i9-7940X CPU @ 3.10GHz

```
$ hyperfine --warmup 3 "~/dev/debian/coreutils/target/release/hashsum --sha256 WinDev2004Eval-disk001.vdi" "sha256sum WinDev2004Eval-disk001.vdi"

Benchmark 1: ~/dev/debian/coreutils/target/release/hashsum --sha256 WinDev2004Eval-disk001.vdi
  Time (mean ± σ):     280.164 s ±  0.878 s    [User: 261.260 s, System: 18.201 s]
  Range (min … max):   278.827 s … 281.291 s    10 runs
 
Benchmark 2: sha256sum WinDev2004Eval-disk001.vdi
  Time (mean ± σ):     231.048 s ±  0.559 s    [User: 214.169 s, System: 16.165 s]
  Range (min … max):   230.556 s … 232.344 s    10 runs
 
Summary
  'sha256sum WinDev2004Eval-disk001.vdi' ran
    1.21 ± 0.00 times faster than '~/dev/debian/coreutils/target/release/hashsum --sha256 WinDev2004Eval-disk001.vdi'
```

the file is 50G 

interesting optim on AMD?!