@useribs 2025-03-15
Interesting, how come Rust implementation can be slower than GNU. Tested on Debian 12
```
root@49131f:/dev/shm# time sha256sum test
612fc818ff21e63f76fb2d45609825aacabab0c1575f978a2c40688990c3b9de  test

real    0m3.069s
user    0m2.980s
sys     0m0.088s
root@49131f:/dev/shm# time /usr/lib/cargo/bin/coreutils/sha256sum ./test
612fc818ff21e63f76fb2d45609825aacabab0c1575f978a2c40688990c3b9de  ./test

real    0m3.490s
user    0m3.363s
sys     0m0.124s
root@49131f:/dev/shm# time sha256sum test
612fc818ff21e63f76fb2d45609825aacabab0c1575f978a2c40688990c3b9de  test

real    0m3.143s
user    0m3.041s
sys     0m0.100s
root@49131f:/dev/shm# time /usr/lib/cargo/bin/coreutils/sha256sum ./test
612fc818ff21e63f76fb2d45609825aacabab0c1575f978a2c40688990c3b9de  ./test

real    0m3.431s
user    0m3.273s
sys     0m0.157s
```