@sylvestre 2025-03-15
@useribs for proper benchmarking, please use hyperfine